Update login API and crendential storage
Recent changes on the CMore side broke the plugin since it was no longer
able to login properly. This change replaces the old login method with
one based on the current Android app, which uses a query for logging in.
The plugin credential handling is also updated to better fit with the
new format of the data returned by the login query. Last, but not least,
a minor update of the header of the stream request was needed, as it
would otherwise not authenticate correctly.